# docs/api.yaml
openapi: 3.0.0
info:
  title: Logbook API
  version: 0.0.1
  description: Logbook API has been designed for internal use only.
servers:
  - url: localhost:8080
...
paths:
...
  /document/list:
    get:
      tags: ["Document"]
      summary: Get list of all documents a user own.
      parameters:
        - name: Bearer
          in: header
          required: true
          schema: { type: string, format: uuid }
      responses:
        200:
          description: Ok.
          content:
            "application/json":
              schema:
                type: object
                items: { type }
...